Create sq files for routes, stops, stoptimes and trips
