+++ feature/sydney-trains/database/api/src/main/sqldelight/xyz/ksharma/krail/sydney/trains/database/stops.sq
@@ -1,0 +1,37 @@
+CREATE TABLE stops (
+    stop_id TEXT PRIMARY KEY,
+    stop_code TEXT,
+    stop_name TEXT,
+    stop_desc TEXT,
+    stop_lat REAL,
+    stop_lon REAL,
+    zone_id TEXT,
+    stop_url TEXT,
+    location_type INTEGER,
+    parent_station TEXT,
+    stop_timezone TEXT,
+    wheelchair_boarding INTEGER
+);
+
+insertIntoStops:
+INSERT OR IGNORE INTO stops (
+    stop_id,
+    stop_code,
+    stop_name,
+    stop_desc,
+    stop_lat,
+    stop_lon,
+    zone_id,
+    stop_url,
+    location_type,
+    parent_station,
+    stop_timezone,
+    wheelchair_boarding
+)
+VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?);
+
+selectAllStops:
+SELECT * FROM stops;
+
+stopsCount:
+SELECT COUNT(*) FROM stops;
