+++ feature/sydney-trains/database/api/src/main/sqldelight/xyz/ksharma/krail/sydney/trains/database/trips.sq
@@ -1,0 +1,32 @@
+CREATE TABLE trips (
+    route_id TEXT NOT NULL,
+    service_id TEXT NOT NULL,
+    trip_id TEXT NOT NULL,
+    trip_headsign TEXT,
+    trip_short_name TEXT,
+    direction_id INTEGER,
+    block_id TEXT,
+    shape_id TEXT,
+    wheelchair_accessible INTEGER,
+    PRIMARY KEY (trip_id)
+);
+
+insertIntoTrips:
+INSERT OR IGNORE INTO trips (
+    route_id,
+    service_id,
+    trip_id,
+    trip_headsign,
+    trip_short_name,
+    direction_id,
+    block_id,
+    shape_id,
+    wheelchair_accessible
+)
+VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?);
+
+selectAllTrips:
+SELECT * FROM trips;
+
+tripsCount:
+SELECT COUNT(*) FROM trips;
