+++ feature/sydney-trains/database/real/src/main/kotlin/xyz/ksharma/krail/sydney/trains/database/real/di/SydneyTrainsDbModule.kt
@@ -4,9 +4,9 @@ import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
-import xyz.ksharma.krail.database.sydney.trains.database.api.SydneyTrainsStaticDB
+import xyz.ksharma.krail.database.sydney.trains.database.api.StopTimesStore
 import xyz.ksharma.krail.database.sydney.trains.database.api.ZipEntryCacheManager
-import xyz.ksharma.krail.sydney.trains.database.real.RealSydneyTrainsStaticDb
+import xyz.ksharma.krail.sydney.trains.database.real.RealStopTimesStore
 import xyz.ksharma.krail.sydney.trains.database.real.RealZipCacheManager
 import javax.inject.Singleton
 
@@ -16,7 +16,7 @@ abstract class SydneyTrainsDbModule {
 
     @Binds
     @Singleton
-    abstract fun bindSydneyTrainsStaticDB(impl: RealSydneyTrainsStaticDb): SydneyTrainsStaticDB
+    abstract fun bindSydneyTrainsStaticDB(impl: RealStopTimesStore): StopTimesStore
 
     @Binds
     @Singleton
